update util function to move more things out of the base app
